* DEL: Cleanup xrGame files.
