--- code/engine.vc2008/xrGame/inventory_owner_inline.h
@@ -1,7 +1,0 @@
-#pragma once
-
-IC	CTradeParameters &CInventoryOwner::trade_parameters	() const
-{
-	VERIFY	(m_trade_parameters);
-	return	(*m_trade_parameters);
-}
