--- code/engine.vc2008/xrGame/xrServer_svclient_validation.cpp
@@ -1,23 +1,0 @@
-#include "stdafx.h"
-#include "xrServer_svclient_validation.h"
-#include "GameObject.h"
-#include "Level.h"
-
-bool is_object_valid_on_svclient(u16 id_entity)
-{
-	CObject* tmp_obj		= Level().Objects.net_Find(id_entity);
-	if (!tmp_obj)
-		return false;
-	
-	CGameObject* tmp_gobj	= smart_cast<CGameObject*>(tmp_obj);
-	if (!tmp_gobj)
-		return false;
-
-	if (tmp_obj->getDestroy())
-		return false;
-
-	if (tmp_gobj->object_removed())
-		return false;
-	
-	return true;
-};
