--- code/engine.vc2008/xrGame/xrServer_svclient_validation.h
@@ -1,2 +1,0 @@
-#pragma once
-bool is_object_valid_on_svclient(u16 id_entity);
